Reorganize AH tags test
